+++ src/main/java/domain/user/Dealer.java
@@ -1,6 +1,7 @@
 package domain.user;
 
 import domain.card.Card;
+import util.Converter;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -11,6 +12,8 @@ import java.util.List;
  */
 public class Dealer {
     private final List<Card> cards = new ArrayList<>();
+
+    private final Converter converter = new Converter();
 
     public Dealer() {}
 
@@ -48,5 +51,13 @@ public class Dealer {
 
     public List<Card> getCards() {
         return cards;
+    }
+
+    public String getCardValues() {
+        List<String> cardValues = new ArrayList<>();
+        for (Card card : cards) {
+            cardValues.add(card.getCardValue());
+        }
+        return converter.convertListToStirng(cardValues);
     }
 }
